app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
bcrypt==4.1.2
flask-socketio==5.3.6
eventlet==0.35.2
Flask-Limiter==3.5.1
Pillow==10.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
numpy
# Dev dependencies
black==24.3.0
flake8==7.0.0
pre-commit==3.6.2
flask_mail
